Refactor math classes to use double precision for improved accuracy in calculations and for handling larger numbers

+++ Engine/Math/headers/BoundingBox.hpp
@@ -13,7 +13,7 @@ namespace Math {
 
         BoundingBox(Vector* min, Vector* max);
 
-        BoundingBox(float minX, float minY, float maxX, float maxY);
+        BoundingBox(double minX, double minY, double maxX, double maxY);
 
         ~BoundingBox();
