+++ Engine/Math/src/Transformation.cpp
@@ -14,12 +14,12 @@ namespace Math {
 
 
 
-    Transformation::Transformation(const Vector& position, float angle) :
+    Transformation::Transformation(const Vector& position, double angle) :
         PositionX(position.x), PositionY(position.y),
         sine(std::sin(angle)), cosine(std::cos(angle)) {
     }
 
-    Transformation::Transformation(float x, float y, float angle) :
+    Transformation::Transformation(double x, double y, double angle) :
         PositionX(x), PositionY(y),
         sine(std::sin(angle)), cosine(std::cos(angle)) {
     }
